Fix eldoc frame shrinking on some setups
There appears to be an Emacs bug where in some environments, calling
set-frame-size to another size and then to the original size will result in the
second call not taking effect. This means that the workaround for #68 causes
the frame to shrink if moving between doc items with the same doc size (or
within the same doc item).

Instead, we just temporarily shrink the Markdown separators to make the window
size calculations work, then restore them afterwards.

* eldoc-box.el (eldoc-box--update-childframe-geometry): Adjust markdown
separator width instead of setting the frame to a tiny size.