Make help-at-point don't show outdated doc (issue#96)
When help-at-point is invoked when eldoc's doc buffer hasn't received
the latest doc, the help-at-point pupup will show the old doc and
never change to the new one.  Now help-at-point will insert a callback
in eldoc-display-functions, so that when any async doc comes in, the
help-at-point popup is updated with the latest doc.

* eldoc-box.el (eldoc-box--help-at-point-async-update): New function.
(eldoc-box-help-at-point): Add hook.