Improve position of child frame for hover at point (issue#134)
When there is not enough space on the right to display the child
frame, move the child frame left so the right edge of the child
frame still in the emacs window, but keep the child frame at
right side of the window as much as possible, so it hide the
code as little as possible. Code is usually concentrated on the
left side of the editor.

* eldoc-box.el:
(eldoc-box--default-at-point-position-function-1): Keep the
childframe on the right.